+++ src/__internal__/input/index.ts
@@ -1,10 +1,2 @@
-import Input from "./input.component";
-import InputPresentation from "./input-presentation.component";
-
-export type {
-  InputProps,
-  CommonInputProps,
-  EnterKeyHintTypes,
-} from "./input.component";
-export { Input, InputPresentation };
-export default Input;
+export { default } from "./input.component";
+export type { InputProps } from "./input.component";