+++ src/__internal__/legacy-error-border/error-border.style.ts
@@ -1,5 +1,9 @@
 import styled, { css } from "styled-components";
 
+/*
+ * This has been left in as some legacy components still rely on it
+ * Once Textarea, Fieldset, FileInput and TextEditor have been aligned this can be deleted
+ */
 const ErrorBorder = styled.span`
   ${({ warning }: { warning: boolean }) => css`
     position: absolute;