+++ src/components/fieldset/fieldset.test.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import { render, screen } from "@testing-library/react";
 import Fieldset from "./fieldset.component";
 import Textbox from "../textbox";
-import CarbonProvider from "../carbon-provider";
 
 test("Fieldset Legend is rendered if supplied", () => {
   render(<Fieldset legend="Legend" />);
@@ -45,19 +44,4 @@ test("Fieldset Legend adds an asterisk after the text when the field is mandator
 
   const legend = screen.getByText("This is my custom legend");
   expect(legend).toBeInTheDocument();
-});
-
-// For coverage, this margin will be different when wrapped in a Form due to default fieldSpacing
-test("renders with default margin between inputs when the validationRedesignOptIn is true", () => {
-  render(
-    <CarbonProvider validationRedesignOptIn>
-      <Fieldset>
-        <Textbox data-role="textbox" value={""} onChange={() => {}} />
-        <Textbox data-role="textbox 2" value={""} onChange={() => {}} />
-      </Fieldset>
-    </CarbonProvider>,
-  );
-
-  const secondInput = screen.getByTestId("textbox 2");
-  expect(secondInput).toHaveStyle("margin-top: 16px");
 });