+++ src/components/numeral-date/__internal__/numeral-date.context.ts
@@ -1,7 +1,0 @@
-import React from "react";
-
-type NumeralContextType = {
-  disableErrorBorder?: boolean;
-};
-
-export default React.createContext<NumeralContextType>({});