+++ CHANGELOG.md
@@ -1,3 +1,97 @@
+## [159.3.0](https://github.com/Sage/carbon/compare/v159.2.5...v159.3.0) (2026-05-15)
+
+### Features
+
+* **required-fields-indicator:** align component with fusion DS ([a1f49c7](https://github.com/Sage/carbon/commit/a1f49c779a8a01eb9c9552120853236e21fee84f))
+
+## [159.2.5](https://github.com/Sage/carbon/compare/v159.2.4...v159.2.5) (2026-05-14)
+
+### Bug Fixes
+
+* **text-editor:** ensure that links can be added after initial render ([8a37110](https://github.com/Sage/carbon/commit/8a37110b6d3bc409f012a0d99595b19f4f0b4dac))
+
+## [159.2.4](https://github.com/Sage/carbon/compare/v159.2.3...v159.2.4) (2026-05-13)
+
+### Bug Fixes
+
+* **use-media-query:** optimise hook to avoid uneccessary rerenders ([f33bec4](https://github.com/Sage/carbon/commit/f33bec42947d6ad2e5c6dd02a3c973571087fd0f)), closes [#7884](https://github.com/Sage/carbon/issues/7884)
+
+## [159.2.3](https://github.com/Sage/carbon/compare/v159.2.2...v159.2.3) (2026-05-13)
+
+### Bug Fixes
+
+* **verticalmenu:** padding fix on trigger ([76ef6e7](https://github.com/Sage/carbon/commit/76ef6e786dbf200f2975848d202a0bbb678b8446))
+
+## [159.2.2](https://github.com/Sage/carbon/compare/v159.2.1...v159.2.2) (2026-05-13)
+
+### Bug Fixes
+
+* **tokens:** update token script to ensure any needed JS tokens are included in build ([bfd99a3](https://github.com/Sage/carbon/commit/bfd99a3559ffed324f90d7ec6ca2c480a5199256))
+
+## [159.2.1](https://github.com/Sage/carbon/compare/v159.2.0...v159.2.1) (2026-05-12)
+
+### Bug Fixes
+
+* **action-popover-item:** ensure enter keydown triggers href redirect ([0c87d7f](https://github.com/Sage/carbon/commit/0c87d7ff0520c4c06be16894375d2c7f311eb50a)), closes [#7945](https://github.com/Sage/carbon/issues/7945)
+
+## [159.2.0](https://github.com/Sage/carbon/compare/v159.1.0...v159.2.0) (2026-05-11)
+
+### Features
+
+* **dialog:** align component with fusion DS ([9e729f0](https://github.com/Sage/carbon/commit/9e729f067b1bafba6534084193084ef56ed78da7))
+
+## [159.1.0](https://github.com/Sage/carbon/compare/v159.0.1...v159.1.0) (2026-05-11)
+
+### Features
+
+* **icon:** added 3 new icons and fixed another ([746685f](https://github.com/Sage/carbon/commit/746685fd594a3a002d62ac3644eebbd2f8c69b4a))
+
+## [159.0.1](https://github.com/Sage/carbon/compare/v159.0.0...v159.0.1) (2026-05-08)
+
+### Bug Fixes
+
+* **checkbox-group:** remove fit-content from group styles ([15b3da5](https://github.com/Sage/carbon/commit/15b3da56f2cae4499f0f905b157785129eea2ba7)), closes [#7915](https://github.com/Sage/carbon/issues/7915)
+
+## [159.0.0](https://github.com/Sage/carbon/compare/v158.50.0...v159.0.0) (2026-05-08)
+
+### ⚠ BREAKING CHANGES
+
+* **textbox:** the legacy interface has been maintained wherever possible, the
+following props no longer have any functionality:
+adaptiveLabelBreakpoint, info, helpAriaLabel, labelAlign,
+labelWidth, labelSpacing, labelId, reverse, tooltipPosition,
+tooltipId, info and validationOnLabel
+(legacy validation has been removed).
+This has been released as a new major version as there is likely
+to be some layout changes when upgrading, these should be solved
+using a parent container to constrain/ position etc.
+
+### Features
+
+* **textbox:** align component with fusion DS ([8360cb3](https://github.com/Sage/carbon/commit/8360cb383d1e0cab9df0a70a016349fea1667b88))
+
+### Bug Fixes
+
+* **numeral-date:** remove fit-content width to allow additional text to overflow the inputs ([54bd887](https://github.com/Sage/carbon/commit/54bd88727b1b7b39eecf11f1edcc902c4599e0fc)), closes [#7875](https://github.com/Sage/carbon/issues/7875)
+
+## [158.50.0](https://github.com/Sage/carbon/compare/v158.49.2...v158.50.0) (2026-05-07)
+
+### Features
+
+* **tabs:** updated to use correct new tokens ([b8f2def](https://github.com/Sage/carbon/commit/b8f2defebf1600bfd22dfc346349c52b5201ac5a))
+
+## [158.49.2](https://github.com/Sage/carbon/compare/v158.49.1...v158.49.2) (2026-05-06)
+
+### Bug Fixes
+
+* **text-editor:** correct WYSIWYG asterisk colour ([d2908ce](https://github.com/Sage/carbon/commit/d2908ce264d89efadce7461d5d261d0ee4cdb450))
+
+## [158.49.1](https://github.com/Sage/carbon/compare/v158.49.0...v158.49.1) (2026-05-05)
+
+### Bug Fixes
+
+* **text-editor:** span from initialValue not being styled ([cd6d041](https://github.com/Sage/carbon/commit/cd6d0417ef1db5fb6027237151637023133a78f3))
+
 ## [158.49.0](https://github.com/Sage/carbon/compare/v158.48.0...v158.49.0) (2026-05-01)
 
 ### Features